+++ src/components/FavouriteCard.tsx
@@ -5,30 +5,24 @@ import {useAppSelector} from "../hooks/redux";
 import {Link} from "react-router-dom";
 
 interface Props {
-    title: string
+    id: string
 }
 
 export function FavouriteCard(prop: Props) {
     const [fetchFilms, {isLoading: loadingFilms, data, isError}] = useLazyGetFilmDescriptionQuery();
 
-    const {addFavourite, removeFavourite} = useActions()
+    const { removeFavourite } = useActions()
     const {favourites} = useAppSelector(state => state.films)
-    const [isFav, setIsFav] = useState(favourites.includes(data?.title!))
-    const addToFavourite = (event: React.MouseEvent<HTMLButtonElement>) => {
-        event.preventDefault()
-        setIsFav(true)
-        addFavourite(data?.id!)
-    }
+    const isFav = favourites.includes(prop.id)
 
-    const removeFromFavourite = (event: React.MouseEvent<HTMLButtonElement>) => {
-        event.preventDefault()
-        setIsFav(false)
+    const removeFromFavourite = () => {
         removeFavourite(data?.id!)
     }
 
     useEffect(() => {
-        fetchFilms(prop.title)
-    }, [prop.title])
+        fetchFilms(prop.id)
+    }, [prop.id])
+
     return (
         // {isLoading && <p>Data is loading</p>}
         <div>
@@ -51,19 +45,11 @@ export function FavouriteCard(prop: Props) {
                         content.
                     </p>
                     <div className="flex justify-content space-x-10 px-4 pt-4">
-                        {isFav && <button onClick={addToFavourite}>
-                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-                                 stroke="#f94144" className="w-6 h-6">
-                                <path stroke-linecap="round" stroke-linejoin="round"
-                                      d="M21 8.25c0-2.485-2.099-4.5-4.688-4.5-1.935 0-3.597 1.126-4.312 2.733-.715-1.607-2.377-2.733-4.313-2.733C5.1 3.75 3 5.765 3 8.25c0 7.22 9 12 9 12s9-4.78 9-12z"/>
-                            </svg>
-                        </button>
-                        }
-                        {!isFav && <button onClick={removeFromFavourite}>
+                        {isFav && <button type="button" onClick={removeFromFavourite}>
                             <svg xmlns="http://www.w3.org/2000/svg" fill="#f94144" viewBox="0 0 24 24"
-                                 stroke-width="1.5"
+                                 strokeWidth="1.5"
                                  stroke="#f94144" className="w-6 h-6">
-                                <path stroke-linecap="round" stroke-linejoin="round"
+                                <path strokeLinecap="round" strokeLinejoin="round"
                                       d="M21 8.25c0-2.485-2.099-4.5-4.688-4.5-1.935 0-3.597 1.126-4.312 2.733-.715-1.607-2.377-2.733-4.313-2.733C5.1 3.75 3 5.765 3 8.25c0 7.22 9 12 9 12s9-4.78 9-12z"/>
                             </svg>
                         </button>
