+++ src/pages/FavouritesPage.tsx
@@ -11,11 +11,12 @@ export function FavouritesPage() {
             <p className="center">No items</p>
         )
     }
+
     return (
 
                 <div className="flex  justify-center flex-wrap flex-initial gap-6 pt-10 mx-auto h-screen w-screen">
                     {favourites.map(f => (
-                        <FavouriteCard key={f} title={f}/>
+                        <FavouriteCard key={f} id={f}/>
                     ))
                     }
                 </div>
