+++ src/store/films/films.slice.ts
@@ -1,10 +1,10 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-const LS_FAV_KEY = 'rfk'
 interface FilmsState {
     favourites: string[]
 }
 
+const LS_FAV_KEY = 'rfk'
 const initialState: FilmsState = {
     favourites: JSON.parse(localStorage.getItem(LS_FAV_KEY) ?? '[]')
 }
@@ -15,11 +15,11 @@ export const filmsSlice = createSlice({
     reducers: {
         addFavourite(state, action: PayloadAction<string>) {
             state.favourites.push(action.payload);
-            localStorage.setItem(LS_FAV_KEY, JSON.stringify(state.favourites));
+            // localStorage.setItem(LS_FAV_KEY, JSON.stringify(state.favourites));
         },
         removeFavourite(state, action: PayloadAction<string>) {
             state.favourites = state.favourites.filter(f => f !== action.payload);
-            localStorage.setItem(LS_FAV_KEY, JSON.stringify(state.favourites))
+            // localStorage.setItem(LS_FAV_KEY, JSON.stringify(state.favourites))
         }
     }
 })
